refactor: separate components LoginPageBtn and ButtonBig
