+++ rs-commerce/src/components/login-page-btn/style.module.css
@@ -1,0 +1,21 @@
+.login-page-btn {
+  border: 1px solid var(--primary);
+  border-radius: 5px;
+  width: 237px;
+  height: 55px;
+  background: var(--primary);
+  font-family: var(--font-poppins);
+  font-weight: 400;
+  font-size: 16px;
+  color: #fff;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+  transition: var(--transition-time);
+}
+.login-page-btn.btn-not-active {
+  border: 1px solid #adadad;
+  background: #adadad;
+  cursor: not-allowed;
+}
